requirements.txt
mysql-connector-python==8.4.0
fastapi==0.111.0
pydantic-settings==2.2.1
pytz==2024.1
setuptools==70.0.0
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.35.0
scipy==1.13.1
plotly==5.22.0